Sample Bots/Java/entelect-battleships/src/test/java/net/avdw/battlefight/hunt/HuntBehaviourTreeTest.java
package net.avdw.battlefight.hunt;

import java.io.File;
import java.io.IOException;
import net.avdw.battlefight.MapQuery;
import net.avdw.battlefight.state.StateModel;
import net.avdw.battlefight.state.StateReader;
import net.avdw.battlefight.struct.Action;
import org.junit.Test;
import static org.junit.Assert.*;

public class HuntBehaviourTreeTest {

    @Test
    public void testCannotFindShip() throws IOException {
        StateModel model = StateReader.read(new File("src/test/resources/cannot-find-ship.json"), StateModel.class);
        StateModel.OpponentCell[][] map = MapQuery.transformMap(model.OpponentMap.Cells);
        MapQuery.printMap(map);
        
        Action action = HuntBehaviourTree.execute(model);
        assertNotEquals("Don't make illegal shots.", "1,0,0", action.toString());
        System.out.println(action);
    }
    
    @Test
    public void testUseEnergy() throws IOException {
        StateModel model = StateReader.read(new File("src/test/resources/cannot-find-ship.json"), StateModel.class);
        
        Action action = HuntBehaviourTree.execute(model);
        assertNotEquals("Should use special shot.", "FIRESHOT", action.type.name());
    }
}
